pyproject.toml
[project]
name = "bilibili-grok"
version = "0.1.0"
description = "AI-powered auto-reply bot for Bilibili @mentions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "qrcode>=7.4",
    "pyyaml>=6.0",
    "litellm>=1.0.0",
    "aiosqlite>=0.19.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3.0",
]

[project.scripts]
grok = "grok.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/grok"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
